(refactor) move agent runtime foundations from handlers/ into condor/ (ARCH-190)
condor/ core no longer imports from handlers/: the platform-neutral pieces
of handlers/agents/_shared.py moved down — MCP toolset builders to
condor/runtime/toolsets.py, danger classification + confirmation summary to
condor/runtime/danger.py, context builders to condor/runtime/context.py,
the model catalog to condor/llm/options.py, TelegramChannel to
condor/runtime/channels.py — and readiness/openrouter_models/custom_models
moved wholesale to condor/llm/. clean_config_for_save moved to
condor/controller_configs.py. handlers modules re-export (or sys.modules-
alias) the moved names so Telegram callers are unchanged; the lazy
function-local imports across condor/runtime, condor/agents and condor/web
are now normal top-level imports, and the trade-gate danger lists no longer
live in a hot-reloaded handlers module.
