+++ condor/agents/consult.py
@@ -26,6 +26,9 @@ from condor.acp.pydantic_ai_client import (
 )
 from condor.agents.agent import AgentStore
 from condor.preferences import resolve_custom_endpoint
+from condor.runtime import context as runtime_context
+from condor.runtime import toolsets
+from condor.runtime.channels import TelegramChannel
 
 log = logging.getLogger(__name__)
 
@@ -42,7 +45,6 @@ def _build_consult_permission_cb(slug: str, user_id: int, chat_id: int):
     try:
         from condor.routine_store import get_routine_store
         from condor.runtime.confirmations import build_permission_callback
-        from handlers.agents.confirmation import TelegramChannel
 
         bot = get_routine_store().get_bot()
         if bot is not None:
@@ -162,11 +164,6 @@ async def _run_agent_to_completion(
 
     # Build the Agent's MCP toolset in the main process (ConfigManager is here).
     # agent_slug scopes the condor MCP tools' memory/skills to this Agent (its brain).
-    from handlers.agents._shared import (
-        build_agent_context,
-        build_mcp_servers_for_session,
-    )
-
     # A server pinned on the Agent itself wins over the ambient chat server; when
     # the agent isn't pinned, fall back to the caller's (chat's) resolved server.
     # Passing server_name=None lets the builder resolve the chat's server.
@@ -174,7 +171,7 @@ async def _run_agent_to_completion(
     # agent_slug the condor MCP tools would target the CHAT's stores.
     effective_server = agent.server_name or server_name
 
-    mcp_servers = build_mcp_servers_for_session(
+    mcp_servers = toolsets.build_mcp_servers_for_session(
         user_id,
         chat_id,
         server_name=effective_server if agent.server_required else None,
@@ -204,7 +201,7 @@ async def _run_agent_to_completion(
         user_id=user_id,
     )
 
-    prompt = build_agent_context(agent, user_id, task, context)
+    prompt = runtime_context.build_agent_context(agent, user_id, task, context)
 
     await client.start()
     try:
