+++ condor/agents/risk.py
@@ -11,6 +11,13 @@ import logging
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any
 
+from condor.runtime.danger import (
+    DANGEROUS_BOT_ACTIONS,
+    is_dangerous_tool_call,
+    tool_call_input,
+    tool_call_name,
+)
+
 if TYPE_CHECKING:
     from .ownership import BotLedger
 
@@ -136,8 +143,6 @@ class RiskEngine:
 
         Returns (allowed, reason).
         """
-        from handlers.agents._shared import tool_call_input
-
         input_data = tool_call_input(tool_call)
         if input_data is None:
             return False, "Tool arguments could not be read"
@@ -186,8 +191,6 @@ class RiskEngine:
 
         Returns (allowed, reason).
         """
-        from handlers.agents._shared import tool_call_input
-
         input_data = tool_call_input(tool_call)
         if input_data is None:
             return False, "Tool arguments could not be read"
@@ -240,12 +243,6 @@ def auto_approve_with_risk_check(
     mistyped one open a live position no session can ever claim. Empty (consults,
     chat, tests) keeps the presence-only check.
     """
-    from handlers.agents._shared import (
-        DANGEROUS_BOT_ACTIONS,
-        is_dangerous_tool_call,
-        tool_call_input,
-        tool_call_name,
-    )
 
     async def callback(tool_call: dict, options: list[dict]) -> dict:
         if is_dangerous_tool_call(tool_call):
