+++ handlers/agents/confirmation.py
@@ -9,13 +9,11 @@ HTTP instead.
 import logging
 from typing import Any
 
-from telegram import Bot, InlineKeyboardButton, InlineKeyboardMarkup
+from telegram import Bot
 
-from condor.runtime.confirmations import (
-    CONFIRMATION_TIMEOUT,
-    PendingConfirmation,
-    get_registry,
-)
+from condor.runtime.channels import TelegramChannel  # noqa: F401  (re-export)
+from condor.runtime.confirmations import CONFIRMATION_TIMEOUT, get_registry
+from condor.runtime.danger import format_tool_summary  # noqa: F401  (re-export)
 
 log = logging.getLogger(__name__)
 
@@ -28,110 +26,9 @@ __all__ = [
 ]
 
 
-def format_tool_summary(tool_call: dict[str, Any]) -> str:
-    """Format a tool call into a human-readable summary for the confirmation message."""
-    from handlers.agents._shared import tool_call_input, tool_call_name
-
-    # The bare name, or an ACP call (``mcp__mcp-hummingbot__manage_bots``) would
-    # match none of the branches below and be approved as an opaque string.
-    tool_name = tool_call_name(tool_call) or "Unknown"
-    input_data = tool_call_input(tool_call)
-    if input_data is None:
-        # The gate sends unreadable calls here on purpose (SEC-093). Say so,
-        # rather than rendering a row of "?" that looks like a parsed call.
-        return f"{tool_name} (arguments could not be read)"
-
-    if tool_name == "place_order":
-        side = input_data.get("trade_type", "?")
-        pair = input_data.get("trading_pair", "?")
-        amount = input_data.get("amount", "?")
-        order_type = input_data.get("order_type", "MARKET")
-        price = input_data.get("price", "")
-        connector = input_data.get("connector_name", "?")
-        summary = f"{side} {amount} {pair} ({order_type})"
-        if price:
-            summary += f" @ {price}"
-        summary += f" on {connector}"
-        return summary
-
-    if tool_name == "manage_executors":
-        action = input_data.get("action", "?")
-        exec_type = input_data.get("executor_type", "")
-        exec_id = input_data.get("executor_id", "")
-        if action == "create" and exec_type:
-            config = input_data.get("executor_config", {})
-            pair = config.get("trading_pair", "?")
-            return f"Create {exec_type} on {pair}"
-        if action == "stop" and exec_id:
-            return f"Stop executor {exec_id[:12]}..."
-        return f"Executor: {action}"
-
-    if tool_name == "manage_bots":
-        action = input_data.get("action", "?")
-        bot_name = input_data.get("bot_name", "?")
-        if action == "deploy":
-            controllers = input_data.get("controllers_config", [])
-            return f"Deploy bot '{bot_name}' with controllers {controllers}"
-        if action == "update_config":
-            config_name = input_data.get("config_name", "?")
-            return f"Update config '{config_name}' on bot '{bot_name}'"
-        return f"Bot '{bot_name}': {action}"
-
-    if tool_name == "manage_gateway_swaps":
-        action = input_data.get("action", "?")
-        pair = input_data.get("trading_pair", "?")
-        side = input_data.get("side", "?")
-        amount = input_data.get("amount", "?")
-        return f"Swap {side} {amount} {pair}"
-
-    if tool_name == "manage_gateway_clmm":
-        action = input_data.get("action", "?")
-        if action == "open_position":
-            return "Open LP position"
-        if action == "close_position":
-            return "Close LP position"
-        return f"CLMM: {action}"
-
-    # Generic fallback
-    return tool_name
-
-
-# Backwards-compatible alias for the pre-registry name.
+# Backwards-compatible alias for the pre-registry name; the implementation
+# moved to condor.runtime.danger (ARCH-190).
 _format_tool_summary = format_tool_summary
-
-
-class TelegramChannel:
-    """Renders a pending confirmation as an inline keyboard in a chat."""
-
-    def __init__(self, bot: Bot, chat_id: int):
-        self._bot = bot
-        self._chat_id = chat_id
-
-    async def deliver(self, pending: PendingConfirmation) -> None:
-        keyboard = InlineKeyboardMarkup(
-            [
-                [
-                    InlineKeyboardButton(
-                        "Approve", callback_data=f"agent:confirm_trade:{pending.id}"
-                    ),
-                    InlineKeyboardButton(
-                        "Reject", callback_data=f"agent:reject_trade:{pending.id}"
-                    ),
-                ]
-            ]
-        )
-        # Naming the asker is not decoration: one chat can talk to several
-        # agents on several servers, and "Approve this action?" on its own does
-        # not say whose action, or where it lands.
-        asked_by = f"\nRequested by {pending.origin}\n" if pending.origin else ""
-        await self._bot.send_message(
-            chat_id=self._chat_id,
-            text=(
-                f"Trade Confirmation\n{asked_by}\n"
-                f"{pending.summary}\n\nApprove this action?"
-            ),
-            reply_markup=keyboard,
-        )
 
 
 async def permission_callback(
