+++ tests/runtime/test_agent_binding.py
@@ -101,7 +101,7 @@ def test_resolve_unbound_session_is_the_default_agent(monkeypatch):
         return [{"name": "condor"}]
 
     monkeypatch.setattr(
-        "handlers.agents._shared.build_mcp_servers_for_session", fake_session_servers
+        "condor.runtime.toolsets.build_mcp_servers_for_session", fake_session_servers
     )
 
     bound = binding.resolve(_spec(agent_key="claude-code"))
@@ -162,7 +162,7 @@ def test_a_missing_default_record_still_starts_the_chat(monkeypatch, tmp_path):
     """An unreadable agents/condor/AGENT.md degrades; it does not fail closed."""
     monkeypatch.setattr(agent_module, "_DATA_ROOT", tmp_path)
     monkeypatch.setattr(
-        "handlers.agents._shared.build_mcp_servers_for_session", lambda *a, **kw: []
+        "condor.runtime.toolsets.build_mcp_servers_for_session", lambda *a, **kw: []
     )
 
     bound = binding.resolve(_spec(agent_key="claude-code"))
@@ -181,7 +181,7 @@ def test_a_named_agent_that_does_not_exist_still_raises(monkeypatch, tmp_path):
 def test_resolve_agent_sets_memory_scope(agents_root, monkeypatch):
     """A bound Agent scopes the MCP subprocess to its own store."""
     monkeypatch.setattr(
-        "handlers.agents._shared.build_mcp_servers_for_session",
+        "condor.runtime.toolsets.build_mcp_servers_for_session",
         lambda *a, **kw: [{"name": "condor", "agent_slug": kw.get("agent_slug")}],
     )
 
@@ -203,7 +203,7 @@ def test_resolve_agent_sets_memory_scope(agents_root, monkeypatch):
 def test_agent_key_precedence(agents_root, monkeypatch):
     """Explicit spec model > Agent's configured model."""
     monkeypatch.setattr(
-        "handlers.agents._shared.build_mcp_servers_for_session", lambda *a, **kw: []
+        "condor.runtime.toolsets.build_mcp_servers_for_session", lambda *a, **kw: []
     )
 
     inherited = binding.resolve(_spec(agent_slug="executor_manager"))
@@ -228,7 +228,7 @@ def test_unknown_agent_slug(agents_root):
 def test_tool_allowlist_enforced(agents_root, registry, monkeypatch):
     """A bound pydantic-ai session is restricted to the Agent's tools."""
     monkeypatch.setattr(
-        "handlers.agents._shared.build_mcp_servers_for_session", lambda *a, **kw: []
+        "condor.runtime.toolsets.build_mcp_servers_for_session", lambda *a, **kw: []
     )
 
     info = asyncio.run(runtime.create_session(_spec(agent_slug="executor_manager")))
@@ -246,7 +246,7 @@ def test_tool_allowlist_enforced(agents_root, registry, monkeypatch):
 def test_bound_session_opens_with_agent_identity(agents_root, registry, monkeypatch):
     """The Agent's own instructions replace the chat assistant's context."""
     monkeypatch.setattr(
-        "handlers.agents._shared.build_mcp_servers_for_session", lambda *a, **kw: []
+        "condor.runtime.toolsets.build_mcp_servers_for_session", lambda *a, **kw: []
     )
 
     asyncio.run(
@@ -262,7 +262,7 @@ def test_bound_session_opens_with_agent_identity(agents_root, registry, monkeypa
 def test_unbound_session_keeps_chat_context(registry, monkeypatch):
     """Without a binding, the assistant context is untouched."""
     monkeypatch.setattr(
-        "handlers.agents._shared.build_mcp_servers_for_session", lambda *a, **kw: []
+        "condor.runtime.toolsets.build_mcp_servers_for_session", lambda *a, **kw: []
     )
 
     asyncio.run(
@@ -278,7 +278,7 @@ def test_unbound_session_keeps_chat_context(registry, monkeypatch):
 def test_switch_destroys_and_respawns(agents_root, registry, monkeypatch):
     """Rebinding stops the old subprocess exactly once and reuses the key."""
     monkeypatch.setattr(
-        "handlers.agents._shared.build_mcp_servers_for_session", lambda *a, **kw: []
+        "condor.runtime.toolsets.build_mcp_servers_for_session", lambda *a, **kw: []
     )
     key = SessionKey.telegram(42)
 
@@ -309,7 +309,7 @@ def test_switch_destroys_and_respawns(agents_root, registry, monkeypatch):
 def test_rebinding_to_same_agent_reuses_session(agents_root, registry, monkeypatch):
     """An empty agent_key must not look like a model change and respawn."""
     monkeypatch.setattr(
-        "handlers.agents._shared.build_mcp_servers_for_session", lambda *a, **kw: []
+        "condor.runtime.toolsets.build_mcp_servers_for_session", lambda *a, **kw: []
     )
 
     async def scenario():
