+++ tests/runtime/test_busy_policy.py
@@ -78,7 +78,7 @@ def registry(tmp_path, monkeypatch):
     monkeypatch.setattr("condor.acp.client.ACPClient", _GatedClient)
     monkeypatch.setattr(session_module, "build_initial_context", lambda *a, **k: "")
     monkeypatch.setattr(
-        "handlers.agents._shared.build_mcp_servers_for_session", lambda *a, **k: []
+        "condor.runtime.toolsets.build_mcp_servers_for_session", lambda *a, **k: []
     )
     return session_module
