+++ tests/runtime/test_session_creation_lock.py
@@ -65,7 +65,7 @@ def registry(monkeypatch):
     monkeypatch.setattr("condor.acp.client.ACPClient", _SlowClient)
     monkeypatch.setattr(session_module, "build_initial_context", lambda *a, **k: "")
     monkeypatch.setattr(
-        "handlers.agents._shared.build_mcp_servers_for_session", lambda *a, **k: []
+        "condor.runtime.toolsets.build_mcp_servers_for_session", lambda *a, **k: []
     )
     return session_module
