+++ tests/runtime/test_session_facade.py
@@ -52,7 +52,7 @@ def registry(monkeypatch):
     # Lives in condor.runtime.binding now; patch it at the source so both the
     # bound and unbound resolution paths see the stub.
     monkeypatch.setattr(
-        "handlers.agents._shared.build_mcp_servers_for_session", lambda *a, **k: []
+        "condor.runtime.toolsets.build_mcp_servers_for_session", lambda *a, **k: []
     )
     return session_module
