+++ tests/test_session_server_access.py
@@ -224,7 +224,7 @@ def registry(monkeypatch):
     monkeypatch.setattr("condor.acp.pydantic_ai_client.PydanticAIClient", _FakeClient)
     monkeypatch.setattr(session_module, "build_initial_context", lambda *a, **k: "")
     monkeypatch.setattr(
-        "handlers.agents._shared.build_mcp_servers_for_session", lambda *a, **k: []
+        "condor.runtime.toolsets.build_mcp_servers_for_session", lambda *a, **k: []
     )
     return session_module
